First GIS based website, GEWA
Some custom features that are included in this webpage that are not in the Web Application platform provided through ESRI and ArcGIS Online are additional NPS logos.  I found the use of these appropriate since all of the original data and shapefiles comes through the Parks Service.  The layer that I chose to use a filter query on is the Natural Vegetation feature layer.  I chose to remove all fields WHERE ‘Wetland’ = ‘estuarine’.  This gives more distinguished boundaries with the Waterbodies and Park Boundary layers; since, these layers overlapped because the water and soil are marshlands at these locations.  I also added in a photo taken of the antique wooden fence line that is a memorable feature of the area.  This also adds context to what the user could experience from visiting the park.